+++ dit-calendar-bot/src/main/kotlin/com/ditcalendar/bot/data/core/Base.kt
@@ -3,8 +3,4 @@ package com.ditcalendar.bot.data.core
 import kotlinx.serialization.Serializable
 
 @Serializable
-abstract class Base {
-    var entryId: Long = 0
-
-    var version: Int = 0
-}
+abstract class Base(var version: Int = 0)
